Do not bash alias ack to ack-grep
vim doesn't start a bash in interactive mode by default, and thus bash
doesn't reads .bashrc and your aliases (which means vim's bash cannot
understand that ack is actually an alias of ack-grep).

It is probably better that you create a symlink ack to ack-grep to
resolve ack's renaming issue on Debian-derived distros.